bvh base code is done

+++ MTRXEngine/MTRXEngine/source/headers/Defs.h
@@ -10,7 +10,7 @@ namespace mtrx
 {
 	#define LOGGER_FILE_SIZE 5242880 // Maximum size of logger file 
 	enum class ColliderType : char { Sphere, Box, Capsule, Mesh, ConvexShape, };	// Type of what a collider is 
-	static const double pi = 3.14159265358;	// The value of pi (close enough)
+	static const float pi = 3.14159265358f;	// The value of pi (close enough)
 	static float gravity = 15.f; // Constant value for gravity (not realistic value but games don't have realistic gravity)
 
 	// Struct that holds some information about a spring
